+++ tests/terraform/checks/resource/kubernetes/example_PrivilegedContainers/main.tf
